refactor: adiciona view
