+++ model/writeCSV.ts
@@ -1,18 +1,7 @@
 import { NewRow } from './interfaceData';
 
-const user_input = require('prompt-sync')({sigint: true})
 import { createObjectCsvWriter as createCsvWriter } from 'csv-writer';
 
-
-export function readNewRow(): NewRow {
-    const new_row = {
-        nome: user_input('Digite o nome: '),
-        peso: user_input('Digite o peso: '),
-        valor: user_input('Digite o valor: '),
-        quant: user_input('Digite a quantidade: ')
-    }
-    return new_row
-}
 
 export async function writeCSV(file_path: string, new_row: NewRow[]): Promise<void> {
     const csv_writer = createCsvWriter({
