+++ service/serviceEstoque.ts
@@ -1,5 +1,5 @@
 import { readCSV } from '../model/readCSV'
-import { writeCSV, readNewRow } from '../model/writeCSV'
+import { writeCSV } from '../model/writeCSV'
 import { NewRow } from '../model/interfaceData'
 
 const file_path = '../db/estoque.csv'
